вернул хелперы в отдельные файлы, теперь туда передается объект i18n
