+++ src/components/Person/Person.tsx
@@ -6,50 +6,9 @@ import { useRouter } from 'next/router'
 import { IPerson } from '@/types/ComponentProps/IMovie'
 import useMediaQuery from '@/hooks/useMediaQuery'
 import { useTranslation } from 'next-i18next'
+import getFilmWord from '@/util/getFilmWord'
 
 const Person = () => {
-    function getFilmWord(number: number | null) {
-        if (number === null) {
-            return
-        }
-
-        let lastTwoDigits = 0
-        if (number >= 10) {
-            lastTwoDigits = +number.toString().slice(-2)
-        }
-
-        if (lastTwoDigits >= 11 && lastTwoDigits <= 14) {
-            return i18n.language === 'en' ? 'movies' : 'фильмов'
-        }
-
-        const stringifiedNum = number.toString()
-        const lastDigit = +stringifiedNum[stringifiedNum.length - 1]
-
-        let filmWord = ''
-        switch (lastDigit) {
-            case 1:
-                filmWord = i18n.language === 'en' ? 'movie' : 'фильм'
-                break
-            case 2:
-            case 3:
-            case 4:
-                filmWord = i18n.language === 'en' ? 'movies' : 'фильма'
-                break
-            case 5:
-            case 6:
-            case 7:
-            case 8:
-            case 9:
-            case 0:
-                filmWord = i18n.language === 'en' ? 'movies' : 'фильмов'
-                break
-            default:
-                break
-        }
-
-        return filmWord
-    }
-
     const isTab = useMediaQuery('(max-width: 1160px)')
     const router = useRouter()
 
@@ -65,7 +24,7 @@ const Person = () => {
     const name = actorData && actorData.name
     const enName = actorData && actorData.enName
     const numberOfFilms = actorData && actorData.movies.length
-    const filmWord = numberOfFilms ? getFilmWord(numberOfFilms) : getFilmWord(null)
+    const filmWord = numberOfFilms ? getFilmWord(numberOfFilms, i18n) : getFilmWord(null)
 
     const sortedMovies = actorData?.movies.sort((a, b) => b.year - a.year)
